Début du nettoyage du code et structuration :
- Supression de quelques commentaires/lignes innutiles
  ainsi que le dossier build qui n'avait rien a faire la

- mise en place d'un .gitignore

- structuration du code en : - src     (pour les .cpp)
			     - include (pour les .h)
			     - main.cpp

- Formattage fait avec clang-format (avec les réglages par défaut)
