+++ .gitignore
@@ -1,0 +1,5 @@
+#Ignorer le dossier de build
+/build
+
+#Ignorer le fichier .vscode
+.vscode/
