--- Algo Construction arbre de décision
@@ -1,29 +1,0 @@
- #lorsque tout les attribut sont discrets
-
- ArbreDecision(T)
-    si "condition d’arret"
-           retourner feuille(T)
-    sinon
-        choisir le "meilleur" attribut i entre 1 et m
-        pour chaque valeur v de l’attribut i
-             T[v] = {(x, y) de T tels que x_i = v}
-             t[v] = ArbreDecision(T[v])
-        fin pour
-        retourner noeud(i, {v -> t[v]})
-    fin si
-
-    #Lorsque tous les attributs sont à valeur dans R
-
- ArbreDecision (T)
-     si " condition d’ arret "
-              retourner feuille (T)
-     sinon
-         choisir le " meilleur " attribut i entre 1 et m
-         choisir une bonne valeur si pour l’ attribut i
-                 Tf = {(x, y) de T tels que xi ≤ si }
-                 T> = {(x, y) de T tels que xi > si }
-                 tf = ArbreDecision (Tf)
-                 t> = ArbreDecision (T>)
-         fin pour
-         retourner noeud (i,si ,v,tf ,t>)
-    fin si
